forge: attribute_namespace fallback, attribute_id
Makes the jinja filter `attribute_namespace` fallback to `other` for
unqualified ids (e.g. `phase` becomes `other.phase`)

This is in line with the jq filter `semconv_attributes`, which does the
same.

Adds symmetric filter `attribute_id` that returns the other part.

This makes it easier to work with unqualified attributes, which is
common for existing Prometheus metrics.