feat: authenticating when using remote archive
Add `auth-token` option to the CLI that allows configuring
authentication when fetching registry from a remote archive.

`auth-token` is passed and handled separately from the virtual directory
path string as the usual `@` character separating userinfo portion
of the URL from the host is reserved for the `refspec`. Taking
the token from the userinfo portion of the URL would be an alternative
to the approach in this commit but would make the parsing a little
bit more cumbersome. I am open to both approach and happy to adjust
the code to whatever the OpenTelemetry team deems better.

Fixes: https://github.com/open-telemetry/weaver/issues/430